Backend/requirements.txt
Flask
Flask-Cors
noisereduce
librosa
soundfile
whisper
contractions
google-generativeai
joblib
